Two fixed 92 and 95
Fixed 92 - Can enter , in the number fields for survey data
Fixed 95 - Shot data should be stored as a double

+++ src/cwCompassExporterCaveTask.cpp
@@ -146,7 +146,7 @@ void cwCompassExportCaveTask::writeChunk(QTextStream& stream, cwSurveyChunk* chu
         cwStationReference from = shot->fromStation();
         cwStationReference to = shot->toStation();
 
-        float shotLength = cwUnits::convert(shot->distance().toDouble(),
+        float shotLength = cwUnits::convert(shot->distance(),
                                             distanceUnit,
                                             cwUnits::Feet);
